# Backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Test API with swagger",
    "version": "1.0.0",
    "description": ""
  },
  "basePath": "/api",
...
  "schemes": [
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
...
    "/user/get/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/notice/get/all": {
      "get": {
        "tags": [
          "Notice"
        ],
        "description": "해당 사용자가 수강하는 모든 과목들에서 최근 3개의 공지사항들을 받아오기",
        "parameters": [],
        "responses": {
          "200": {
            "description": "정상적으로 공지사항들을 모두 받아온 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "notices": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "_id": {
                        "type": "number",
                        "example": 0
                      },
                      "subject": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "string",
                            "example": "캡스톤디자인"
                          },
                          "professor": {
                            "type": "number",
                            "example": 0
                          },
                          "start_period": {
                            "type": "string",
                            "example": "2021-03-02"
                          },
                          "end_period": {
                            "type": "string",
                            "example": "2021-06-30"
                          },
                          "start_time": {
                            "type": "array",
                            "example": [
                              "16:30",
                              "18:00",
                              "19:30"
                            ],
                            "items": {
                              "type": "string"
                            }
                          },
                          "end_time": {
                            "type": "array",
                            "example": [
                              "18:00",
                              "19:30",
                              "21:00"
                            ],
                            "items": {
                              "type": "string"
                            }
                          },
                          "days": {
                            "type": "array",
                            "example": [
                              1,
                              1,
                              1
                            ],
                            "items": {
                              "type": "number"
                            }
                          },
                          "code": {
                            "type": "string",
                            "example": "519hi32hkjifb12"
                          },
                          "lectures": {
                            "type": "array",
                            "example": [],
                            "items": {}
                          },
                          "students": {
                            "type": "array",
                            "example": [
                              0
                            ],
                            "items": {
                              "type": "number"
                            }
                          },
                          "introURL": {
                            "type": "string",
                            "example": ""
                          }
                        }
                      },
                      "title": {
                        "type": "string",
                        "example": "중간발표 5월 20일 ㅋㅋ"
                      },
                      "content": {
                        "type": "string",
                        "example": "제발 동영상만 내세요! PPT 필요 없습니다!"
                      },
                      "date": {
                        "type": "string",
                        "example": "2021-05-05T15:38:19.424Z"
                      },
                      "comments": {
                        "type": "array",
                        "example": [],
                        "items": {}
                      },
                      "emotions": {
                        "type": "array",
                        "example": [],
                        "items": {}
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/notice/get/subject/{id}": {
      "get": {
        "tags": [
          "Notice"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "정상적으로 해당 subject에 포함된 모든 공지사항들을 받아온 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "notices": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "example": 0
                      },
                      "subject": {
                        "type": "number",
                        "example": 0
                      },
                      "title": {
                        "type": "string",
                        "example": "중간발표 5월 20일 ㅋㅋ"
                      },
                      "content": {
                        "type": "string",
                        "example": "제발 동영상만 내세요! PPT 필요 없습니다!"
                      },
                      "date": {
                        "type": "string",
                        "example": "2021-05-05T15:38:19.424Z"
                      },
                      "comments": {
                        "type": "array",
                        "example": [],
                        "items": {}
                      },
                      "emotions": {
                        "type": "array",
                        "example": [],
                        "items": {}
                      }
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/notice/get/{id}": {
      "get": {
        "tags": [
          "Notice"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "정상적으로 해당 id의 공지사항을 받아온 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "notice": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "number",
                      "example": 0
                    },
                    "subject": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "number",
                          "example": 0
                        },
                        "name": {
                          "type": "string",
                          "example": "캡스톤디자인"
                        },
                        "start_period": {
                          "type": "string",
                          "example": "2021-03-02"
                        },
                        "end_period": {
                          "type": "string",
                          "example": "2021-06-30"
                        },
                        "start_time": {
                          "type": "array",
                          "example": [
                            "16:30",
                            "18:00",
                            "19:30"
                          ],
                          "items": {
                            "type": "string"
                          }
                        },
                        "end_time": {
                          "type": "array",
                          "example": [
                            "18:00",
                            "19:30",
                            "21:00"
                          ],
                          "items": {
                            "type": "string"
                          }
                        },
                        "days": {
                          "type": "array",
                          "example": [
                            1,
                            1,
                            1
                          ],
                          "items": {
                            "type": "number"
                          }
                        },
                        "code": {
                          "type": "string",
                          "example": "519hi32hkjifb12"
                        },
                        "lectures": {
                          "type": "array",
                          "example": [],
                          "items": {}
                        },
                        "students": {
                          "type": "array",
                          "example": [
                            0
                          ],
                          "items": {
                            "type": "number"
                          }
                        },
                        "introURL": {
                          "type": "string",
                          "example": ""
                        }
                      }
                    },
                    "title": {
                      "type": "string",
                      "example": "중간발표 5월 20일 ㅋㅋ"
                    },
                    "content": {
                      "type": "string",
                      "example": "제발 동영상만 내세요! PPT 필요 없습니다!"
                    },
                    "date": {
                      "type": "string",
                      "example": "2021-05-05T15:38:19.424Z"
                    },
                    "comments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "user": {
                            "type": "object",
                            "properties": {
                              "_id": {
                                "type": "number",
                                "example": 0
                              },
                              "name": {
                                "type": "string",
                                "example": "김민건"
                              },
                              "email": {
                                "type": "string",
                                "example": "kkimbj18@ajou.ac.kr"
                              },
                              "photourl": {
                                "type": "string",
                                "example": ""
                              },
                              "school": {
                                "type": "string",
                                "example": "아주대학교"
                              },
                              "identityID": {
                                "type": "string",
                                "example": "201620912"
                              },
                              "major": {
                                "type": "string",
                                "example": "소프트웨어학과"
                              },
                              "type": {
                                "type": "string",
                                "example": "student"
                              },
                              "subject": {
                                "type": "array",
                                "example": [
                                  0,
                                  1,
                                  2,
                                  3
                                ],
                                "items": {
                                  "type": "number"
                                }
                              }
                            }
                          },
                          "content": {
                            "type": "string",
                            "example": "오 쉣 ㅋㅋㅋㅋ 무냐구~~~~"
                          },
                          "date": {
                            "type": "string",
                            "example": "2021-05-05T15:38:19.424Z"
                          }
                        }
                      }
                    },
                    "emotions": {
                      "type": "array",
                      "example": [],
                      "items": {}
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "404": {
            "description": "해당하는 공지가 존재하지 않을 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "existNotice": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/notice/create": {
      "post": {
        "tags": [
          "Notice"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "subject": {
                  "type": "number",
                  "example": 0
                },
                "title": {
                  "type": "string",
                  "example": "오늘은 여기까지만..."
                },
                "content": {
                  "type": "string",
                  "example": "힘들드아"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "정상적으로 공지를 생성했을 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "notice": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "number",
                      "example": 0
                    },
                    "subject": {
                      "type": "number",
                      "example": 0
                    },
                    "title": {
                      "type": "string",
                      "example": "오늘은 여기까지만..."
                    },
                    "content": {
                      "type": "string",
                      "example": "힘들드아"
                    },
                    "date": {
                      "type": "string",
                      "example": "2021-05-05T15:38:19.424Z"
                    },
                    "comments": {
                      "type": "array",
                      "example": [],
                      "items": {}
                    },
                    "emotions": {
                      "type": "array",
                      "example": [],
                      "items": {}
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "403": {
            "description": "type이 professor가 아닌 경우",
            "schema": {
              "$ref": "#/definitions/proAuthFailed"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/notice/update": {
      "put": {
        "tags": [
          "Notice"
        ],
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "type": "object",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 0
                },
                "title": {
                  "type": "string",
                  "example": "오늘은 여기까지만..."
                },
                "content": {
                  "type": "string",
                  "example": "힘들드아"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "정상적으로 공지를 수정했을 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": true
                },
                "notice": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "number",
                      "example": 0
                    },
                    "subject": {
                      "type": "number",
                      "example": 0
                    },
                    "title": {
                      "type": "string",
                      "example": "오늘은 여기까지만..."
                    },
                    "content": {
                      "type": "string",
                      "example": "힘들드아"
                    },
                    "date": {
                      "type": "string",
                      "example": "2021-05-05T15:38:19.424Z"
                    },
                    "comments": {
                      "type": "array",
                      "example": [],
                      "items": {}
                    },
                    "emotions": {
                      "type": "array",
                      "example": [],
                      "items": {}
                    }
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "401": {
            "description": "user가 로그인이 되지 않은 경우",
            "schema": {
              "$ref": "#/definitions/authFailed"
            }
          },
          "403": {
            "description": "type이 professor가 아닌 경우",
            "schema": {
              "$ref": "#/definitions/proAuthFailed"
            }
          },
          "404": {
            "description": "해당 notice가 존재하지 않는 경우",
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "existNotice": {
                  "type": "boolean",
                  "example": false
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "authFailed": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "isLogined": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "proAuthFailed": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": false
        },
        "isProfessor": {
          "type": "boolean",
          "example": false
        }
      }
    }
  }
}